My_Mart_Backend/requirements.txt
flask
flask-pymongo
flask-jwt-extended
flask-bcrypt
python-dotenv
